+++ wazimap_ng/general/widgets.py
