+++ wazimap_ng/profile/admin/admins/profile_highlight_admin.py
@@ -4,7 +4,7 @@ from adminsortable2.admin import SortableAdminMixin
 from ... import models
 from ..forms import ProfileHighlightForm
 
-from wazimap_ng.admin_utils import customTitledFilter, description
+from wazimap_ng.general.widgets import customTitledFilter, description
 from wazimap_ng.datasets.models import Indicator, Dataset
 from wazimap_ng.general.services import permissions
 from wazimap_ng.general.admin.admin_base import BaseAdminModel
