+++ wazimap_ng/profile/admin/forms/profile_highlight_form.py
@@ -4,7 +4,7 @@ from django import forms
 
 from ... import models
 from wazimap_ng.datasets.models import Indicator
-from wazimap_ng.admin_utils import VariableFilterWidget
+from wazimap_ng.general.widgets import VariableFilterWidget
 
 logger = logging.getLogger(__name__)
